### 2017-11-20 #467 @whedon
[PRE REVIEW]: Bitstream - Binary Data for Humans
**Submitting author:** @boisgera (<a href="http://orcid.org/0000-0003-4685-8099">Sébastien Boisgérault</a>)
**Repository:** <a href="https://github.com/boisgera/bitstream" target ="_blank">https://github.com/boisgera/bitstream</a>
**Version:** v2.2.0
