### 2017-11-21 #467 @labarba
@boisgera: I read the paper and it does not illuminate the research application, I'm afraid. Could you elaborate on the research that this software enables, or if it has been used in any context where a scholarly publication would appropriately need to cite its use?
