### 2017-11-22 #467 @boisgera
@labarba: the main selling point of bitstream in the context of research is the high-level API that makes prototyping and experimenting new data compression schemes much easier/faster. However so far it has been us mostly for expository purposes in graduate courses, to replicate and demonstrate the internals of existing formats and codecs, not for any published research yet, but I believe that the potential is here. I mentioned the audio.shrink project since I believe that it's a better proof of concept for research (with a new lossless audio codec with 6 different modes of increasing complexity and efficiency). But this project itself is not documented ... would that help if it was?
