### 2017-11-22 #467 @labarba
Well, you have to explain in the paper what is the research application for the software. As indicated in the [Author Guidelines](http://joss.theoj.org/about#author_guidelines), this explanation is targeted at non-specialist readers and makes a clear statement of need. I don't have familiarity with the topic of data compression, so to look for an appropriate reviewer for your submission, I will need that explanation.
