### 2017-11-22 #467 @boisgera
Fair enough! I will update the paper asap to communicate more clearly to the non-specialist what problem the library solves in data compression research. Thank you for your feedback.
