### 2017-11-25 #467 @boisgera
@labarba I have updated the paper with an explanation and statement of needs that target the non-specialist. A few colleagues (not in the field) that have read both versions considered this one more palatable. I hope it helps; please let me know if I can do something else.
