### 2018-01-22 #467 @labarba
@bmcfee — Would you be interested in being a reviewer for this JOSS submission? (I can send more information on the process by email, if you've not reviewed for us before.)
