### 2018-01-22 #467 @bmcfee
@labarba sure, thanks for the invite!  I haven't reviewed for JOSS before, but I have published in JOSS so I'm familiar with the process.  (Official instructions are always appreciated though.)
