### 2018-01-22 #541 @labarba
Thanks for agreeing to review, @bmcfee. This is where the action happens. Go over the review checklist above, add your comments for the author here, and open issues in the software repo as needed (linking to them in your comments here).
