### 2018-01-22 #541 @bmcfee
Comments on the install procedure:

The instructions in the [docs](https://boisgera.github.io/bitstream/installation/) don't quite get all the way there.  The usual `python setup.py install` or `pip install -e .` failed for me, but the script gave an error message saying to use `python setup.py --cython install`, which did work.  It would be helpful if the install docs had a section for building from source, as opposed to installing from pip.
