### 2018-01-22 #541 @bmcfee
Comments on the documentation:

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

This part seems incomplete.  The documentations state the functionality of the package, but not it's target applications or audience.  I get that it provides a fairly low-level operation, but it would be helpful to mention why someone might want to encode data as a bit sequence, and how to use the data coming out.

The section `Snapshots`, which is intended to demonstrate the `State` object's functionality, actually provides some nice example applications of reading wave files and gene sequences.  Maybe these could be broken out or linked up in the overview?

>  Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

Having now read through the docs, I noticed that the section I requested above (install from source) exists within the contributing section.  Maybe it should be moved or cross-referenced from the install section?

>  Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

The docstrings provide plenty of simple usage examples, but no real-world analysis problems.

>  Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

Mostly:
- The class doc for `BitStream` would be improved (with a bit of redundancy) to list out the constructor parameters.  As is, the reader has to look up `BitStream.write()` to see how to construct a `BitStream`.  Since the call signature is quite small (two params), this amount of redundancy seems acceptable.
- The docstring for `BitStream.read` doesn't specify the `type` of returned values:
```
data: n items of data

The type of data depends on type and n.
```
Is `data` a list of `self.type` entries?  Some other iterable?

- The `State` class documentation is basically empty.

>  Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

Yes -- tests run through doctest and run through travis CI.

>  Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Yes to 1 and 2, but no explicit mention of support.
