### 2018-01-22 #541 @bmcfee
The paper looks pretty good.  On the last point

> References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

Nothing in the bibliography lists a DOI.
