### 2018-01-22 boisgera/bitstream#4 @bmcfee
[JOSS Review] Install from source documentation
Hello!  I'm reviewing your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/541), and will be creating a handful of issues to keep tabs on points in the review form.

The instructions to install from source are listed in the `contributing` section, but I think most people would expect to see this (also) in the `installation` section.
