### 2018-01-22 boisgera/bitstream#5 @bmcfee
[JOSS Review] Provide motivation and example applications in the documentation
[Reference thread https://github.com/openjournals/joss-reviews/issues/541#issuecomment-359456428]

This issue bundles a few points that come up across the JOSS review template.  The paper does a good job of motivating the package and explaining its intended use cases, but this is largely missing from the documentation / overview.  Someone coming across the package without having read the JOSS paper might not get a good sense of what the package can be used for.

Here are a couple of suggestions to improve the docs:

- Reproduce some of the motivation from the JOSS paper in the docs.  This need not be anything big, a couple of sentences could go a long way.
- Make an example applications section.  The `Snapshots` section provides some great examples (wav and gene sequences) that I wouldn't have found if not for reading through the docs in order.  If there was an explicit example applications section, these would be easier to find, and you could link directly to them in the overview to help readers understand the applications better.
