### 2018-01-22 boisgera/bitstream#6 @bmcfee
[JOSS Review] List dependencies in install scripts
[Reference thread https://github.com/openjournals/joss-reviews/issues/541#issuecomment-359456428]

The setup script does not currently list the requirements for installation, runtime, or testing, so users have to install these explicitly.  It would be better to have `requires` and `extras_require` sections of the setup to automate this process, and handle optional dependencies for testing and doc building.
