### 2018-01-22 boisgera/bitstream#7 @bmcfee
[JOSS Review] Improve API docs for State object
[Reference https://github.com/openjournals/joss-reviews/issues/541#issuecomment-359456428]

The documentation for the `State` object does not provide much in the way of detail.  I understand that this is primarily an internal-facing class, but it should still be documented for the sake of future developers.
