### 2018-01-22 boisgera/bitstream#8 @bmcfee
[JOSS Review] Specify return type for stream reading
[Reference https://github.com/openjournals/joss-reviews/issues/541#issuecomment-359456428]

The `BitStream.read` method documentation does not specify what type of object `data` is.  Presumably it's a `list` of `type`s?  Please clarify this point.
