### 2018-01-22 boisgera/bitstream#9 @bmcfee
[JOSS Review] BitStream constructor documentation
[Reference https://github.com/openjournals/joss-reviews/issues/541#issuecomment-359456428]

The `BitStream` constructor documentation refers the reader to the `write` method to learn what `*args, **kwargs` parameters are accepted.  Since the `write` signature is small, it would be easier to read if the constructor was described explicitly.
