### 2018-01-22 boisgera/bitstream#10 @bmcfee
[JOSS Review] Add DOIs for references
[Reference https://github.com/openjournals/joss-reviews/issues/541#issuecomment-359456428]

The `References` section of `paper.md` does not currently provide DOIs for cited items.  Please either add them, or mark here each one that lacks a DOI.
