### 2018-01-22 boisgera/bitstream#11 @bmcfee
[JOSS Review] Community guidelines
[Ref https://github.com/openjournals/joss-reviews/issues/541#issuecomment-359456428]

The contributing section of the docs is pretty good.  I have a couple of minor suggestions though:

- Add an explicit section for people seeking support.  This is distinct from contributing, and should be easy to find.
- Consider copying the contributing doc as `CONTRIBUTING.md` on github.  [This blog post](https://github.com/blog/1184-contributing-guidelines) describes that GH will find and link this directly.
