### 2018-01-24 133246d
document 'read' method return type
