### 2018-01-24 boisgera/bitstream#8 @boisgera
That's the gist of it, yes. But the details are a bit more complex:
it doesn't always make sense to use a list as a container. Here is the rationale:

  - if you call say `stream.read(bool)`, you *probably* don't want a list of one bool,
    but one bool, so that's what we return. If you actually want a list, you can
    use `stream.read(bool,n=1)` instead. So the return type depends on `n`
    (more precisely on `n` being `None` or not).

  - for "container types" (`str`, `BitStream`), you don't want a list either.
    For example `stream.read(str, 10)` shall return a string of length 10,
    not a list of 10 strings of length 1.

  - when you deal with fixed-sized integers (or floating-point numbers),
    especially a large number of them, it makes more sense to use
    NumPy arrays than lists.

  - of course if you use custom types (that are user-defined), all bets are off.
    The documentation of the custom type reader should provide the information.

The "Built-in types" and "Custom types" sections of the documentation discuss this
and are referred to in the "See Also" section of the method,
but I agree that it's not clear (or fast enough ...).

For quick access, I have added a table in the documentation of `read` method
that provides the type of `data` for builtin types (133246d2586c42228e31d4e6643753bc3ad2d0ea).
I guess that it was the kind of info you were searching for.

Does it work for you?
