### 2018-01-24 87f25ea
improve constructor documentation (#9)
