### 2018-01-24 boisgera/bitstream#9 @boisgera
Agreed, not convenient. I have also split the constructor documentation in two: 87f25ea63edf583aa3a4e5be4685bff3717dd12c
