### 2018-01-24 0a84d53
improve doc for State (#7)
