### 2018-01-24 boisgera/bitstream#7 @boisgera
The absence of API is a feature :) ... since this is an [opaque type](https://en.wikipedia.org/wiki/Opaque_data_type). In other words, the *user* should not rely on any implementation detail to be safe: it can only produce one with `save`, store it somehow and if needed give it to `restore`.

That being said:

  - the mere mention of  "(opaque) type" in the documentation is probably not good enough for the user.
    I have added a link to wikipedia, let me know if you think that I should be more explicit.

  - mostly for the developers, I have added a (sketchy) discussion of the internals in the Snapshot section: <http://boisgera.github.io/bitstream/snapshots/#how-does-it-work>. Does it help?

Commit: 0a84d53119f243648ae08a37ab0fce58bab302e3
