### 2018-01-24 boisgera/bitstream#4 @boisgera
Hello Brian! Thanks for taking the time to do this!
I'll have a look at your issues asap.

For the installation section, you're right, that's certainly where I would look for such instructions.
The rationale here was to avoid exposing the user to Cython if it was not needed (I can't help the developer here ...). But I may as well transfer the section with the appropriate warnings.

OK, that's on my TODO list.
