### 2018-01-24 2c59c39
reorganized install vs contribute docs (#4)
