+++ mkdocs/contributing.md
@@ -15,52 +15,16 @@ The following sections should help you to get started.
 
 -----
 
-In the sequel we assume that you are in the project directory:
-
-    $ git clone https://github.com/boisgera/bitstream
-    $ cd bitstream
-
-Build from Sources
+Getting Started
 --------------------------------------------------------------------------------
 
-Bitstream is a [Cython] project.
-Its source files have `.pyx` and `.pxd` extensions. 
-
-To build and install bitstream from its git repository:
-
-  - [install the tools required for a classic install](installation),
-
-  - [install the Cython compiler][Cython],
-
-then execute:
-
-    $ python setup.py --cython install
-
-??? warning "The command `python setup.py install` doesn't work"
-
-    This is expected:
-    this command needs C files to build bitstream,
-    which is a Python C extension module,
-    but these files are not present in the git repository.
-    They have first to be generated by the Cython compiler
-    from the `.pyx` and `.pxd` files with:
-
-        $ python setup.py --cython build
-
-    To "solve" this issue, you can make the `--cython` option implicit
-    with a `setup.cfg` file that contains:
-
-        [global]
-        cython = 1
-
-[Cython]: http://cython.org/#documentation
-[install Cython]: http://cython.readthedocs.io/en/latest/src/quickstart/install.html#installing-cython
-
+First, follow the steps required to [install bitstream from sources](installation/#install-from-sources).
+Make sure that you are in the project directory.
 
 Run the Tests
 --------------------------------------------------------------------------------
 
-To run all the tests, install `pyyaml`[^1] then install bitstream and type
+To run all the tests, install `pyyaml`[^1] and type
 
      $ python test.py
 
