+++ mkdocs/installation.md
@@ -2,7 +2,12 @@
 
 Bitstream supports Python 2.7 on Linux, Windows and MacOS.
 
-Installation
+We recommend you to install bitstream with pip. 
+If for some reason that doesn't work for you 
+-- for example if you want to modify the software -- 
+you can also install bitstream from sources.
+
+Install with Pip
 --------------------------------------------------------------------------------
 
 Check the following prerequisites 
@@ -36,6 +41,39 @@ then install bitstream:
 
 
 
+
+Install from Sources
+--------------------------------------------------------------------------------
+
+Bitstream is a [Cython] project.
+Its source files have `.pyx` and `.pxd` extensions.
+
+??? note "Download the Sources"
+
+    For example, with [git](https://git-scm.com/):
+
+        $ git clone https://github.com/boisgera/bitstream
+        $ cd bitstream
+
+??? note "Pip Install Dependencies"
+
+    [Install the tools required for a pip install](#install-with-pip).
+
+??? note "Cython" 
+
+    [Install the Cython compiler][Installing Cython]. Typically:
+
+        $ pip install Cython
+
+You're now ready to install bitstream: run the command
+
+        $ python setup.py --cython install
+
+[Cython]: http://cython.org/#documentation
+[Installing Cython]: http://docs.cython.org/en/latest/src/quickstart/install.html
+[install Cython]: http://cython.readthedocs.io/en/latest/src/quickstart/install.html#installing-cython
+
+
 Troubleshooting
 --------------------------------------------------------------------------------
 
@@ -61,5 +99,23 @@ Troubleshooting
     and finally install bitstream
 
         $ python -m pip install bitstream
+
+
+??? warning "The command `python setup.py install` doesn't work"
+
+    This is expected:
+    this command needs C files to build bitstream,
+    which is a Python C extension module,
+    but these files are not present in the git repository.
+    They have first to be generated by the Cython compiler
+    from the `.pyx` and `.pxd` files with:
+
+        $ python setup.py --cython build
+
+    To "solve" this issue, you can make the `--cython` option implicit
+    with a `setup.cfg` file that contains:
+
+        [global]
+        cython = 1
 
 
