### 2018-01-24 boisgera/bitstream#4 @boisgera
Commit: 2c59c399c6a4d093c6b6cc80e9cc37c8dc4f7312
