### 2018-01-24 boisgera/bitstream#6 @boisgera
Yes, I do not rely on the setup script for dependencies ... but I wish I could :)

AFAICT:

 1. using `extras_require` to deal with pyyaml (needed for the tests) and mkdocs (to build the docs) should be straightforward (say with an extra called `dev`).

 2. numpy used to be a giant pain to include as a dependency in setup.py. Nowadays, it is much easier (wheels, yay !) ... but **as long as you don't need its C headers, which we do**. Last time I looked, there were unofficial hacks floating around on forums doing things in setup.py that I couldn't understand. If you know better, please tell me, I would need help on this one.

 3. Cython is similar to numpy in this respect: you need cython early in the setup.py (in imports), when the setup.py declared dependencies have not kicked in yet. So again, there are hacks (try/catch the import, replace with a dummy module, wait the setup.py to be reexecuted later when Cython is finally available) but nothing official and I can't tell how robust these methods are. Same here: if you can contribute on this point you're welcome :)

Given that I didn't know how to automate the whole process, I didn't bother doing "the right thing" with the easy part (point 1.: pyyaml and mkdocs). Do you feel that it's worth it nonetheless given the big picture? I'll have a look if you think so.
