### 2018-01-24 boisgera/bitstream#11 @boisgera
- Yeah, good one; I'll add a (short) section for people seeking support directly in the README, so that they don't have to search in the documentation.

  - I knew about CONTRIBUTING.md but I am not a big fan of it. If I could bypass the "GitHub way" and directly declare the url of the Contributing section in the docs, and that link would be displayed when a user writes an issue, that would be great. But here I would have to duplicate the file (urk ...), convert between different markdown flavors (urk ...), rewrite the relative links, etc.
