### 2018-01-24 boisgera/bitstream#6 @bmcfee
Yeah, I've been through a lot of the same hassles, and eventually gave up on cython because of it.

Maybe a decent compromise would be to have a `requirements.txt` file instead of using setup.py?  That way, new users don't need to look up the individual dependencies, and just say `pip install -r requirements.txt` before running setup.  It's not a perfect solution, but it's better than nothing.
