### 2018-01-24 boisgera/bitstream#5 @boisgera
Agreed (both points), will do.
