### 2018-01-24 306ee2b
add DOI for bitstream (#10)
