+++ paper/bibliography.yaml
@@ -11,6 +11,7 @@ references:
   issued:
   - year: 2017
   URL: https://github.com/boisgera/bitstream
+  DOI: 10.5281/zenodo.1158626
 
 - id: DAC
   type: webpage
