### 2018-01-24 boisgera/bitstream#10 @boisgera
AFAICT, only the bitstream refernce has a DOI (via zenodo), I have added it.

Commit: 306ee2b1aba706995e1aa74f2b32f10404ee3f7f
