### 2018-01-24 #541 @boisgera
Regenerate pdf paper (added DOI for bitstream reference)
