### 2018-01-24 ef0eb81
Spoon example (#5)
