### 2018-01-25 9ea00cb
Wave example (#5)
