+++ mkdocs/examples.md
@@ -1,0 +1,238 @@
+
+Examples
+================================================================================
+
+
+
+    >>> from bitstream import BitStream
+    
+
+
+Spoon
+--------------------------------------------------------------------------------
+
+[Spoon] is a derivative of the [Brainfuck] programming language.
+Instead of the 8 ASCII symbols used by Brainfuck, it relies on
+binary sequences to represent each instruction. 
+
+[Spoon]: https://esolangs.org/wiki/Spoon
+[Brainfuck]: https://en.wikipedia.org/wiki/Brainfuck
+
+The source of the "Hello World!" program in Brainfuck is[^dtm]:
+
+    >>> hello_world = "++++++++++[>+++++++>++++++++++>+++>+<<<<-]>++.>+.+++++++..+++.>++.<<+++++++++++++++.>.+++.------.--------.>+.>."
+
+[^dtm]: You don't have to trust me on this, you may try this code online at <https://copy.sh/brainfuck/>.
+
+The conversion between Brainfuck and Spoon is given by the table:
+
+    >>> spoon = {
+    ...   "+": "1",
+    ...   "-": "000",
+    ...   ">": "010",
+    ...   "<": "011",
+    ...   "[": "00100",
+    ...   "]": "0011",
+    ...   ".": "001010",
+    ...   ",": "0010110",
+    ... }
+
+You may notice that the binary codes that correspond to symbols that frequently
+appear in the `hello_world` program (such as `+`) are shorter than the others,
+a nice property since it generates a compact representation of Spoon programs[^nac].
+
+[^nac]: This is not a coincidence: Spoon has been designed as 
+a [Huffman code](https://en.wikipedia.org/wiki/Huffman_coding)
+based on the analysis of a representative collection of Brainfuck programs.
+
+To create a bitstream that contains the Spoon translation of `hello_world`,
+we may for each symbol in this program find the corresponding binary code 
+in the `spoon` table, transform it into a list of bools and append this data 
+to the stream.
+
+    >>> stream = BitStream()
+    >>> for symbol in hello_world:
+    ...     code = spoon[symbol]
+    ...     code_as_bools = [bool(int(char)) for char in code]
+    ...     stream.write(code_as_bools)
+
+Here is the result:
+
+    >>> stream
+    11111111110010001011111110101111111111010111010101101101101100000110101100101001010010101111111001010001010111001010010110010100110111111111111111110010100100010101110010100000000000000000000010100000000000000000000000000010100101001010010001010
+    >>> len(hello_world) * 8
+    888
+    >>> len(stream)
+    245
+
+We now have of program of 245 bits instead of the 888 bits 
+(111 bytes) of the original ASCII program. Not bad ...
+
+-----
+
+It's also pretty easy to perform the opposite operation, to translate Spoon into
+Brainfuck. First, we can compute `noops`, the inverse of the `spoon` dictionary.
+
+    >>> noops = {}
+    >>> for symbol, code in spoon.items():
+    ...     noops[code] = symbol
+    >>> n = max(len(code) for code in noops)
+
+Then we read bits one by one from the stream into a buffer and look at each 
+stage if the buffer corresponds to a key in `noops`, translate this code,
+empty the buffer and start again[^pfc].
+
+    >>> src = ""
+    >>> buffer = BitStream()
+    >>> while len(stream) > 0:
+    ...     if len(buffer) > n:
+    ...         raise ValueError("invalid Spoon bitstream")
+    ...     else:
+    ...         buffer.write(stream.read(bool))
+    ...         try:
+    ...             src += noops[str(buffer)]
+    ...             buffer = BitStream()
+    ...         except KeyError:
[…]
