### 2018-01-25 c23434a
added WAVE example hash test
