### 2018-01-25 ecbc546
doc: extra paragraph
