### 2018-01-25 b963d62
examples: introduction (#5)
