### 2018-01-25 20e9710
WAVE example: forgotten paragraph (#5)
