### 2018-01-25 f971412
test deps via requirements.txt (#6)
