### 2018-01-25 a9ed614
requirements.txt tweak (#6)
