+++ .travis.yml
@@ -5,10 +5,10 @@ language: bash
 before_install:
   - curl https://repo.continuum.io/miniconda/Miniconda2-latest-Linux-x86_64.sh > conda.sh
   - bash conda.sh -b -p $HOME/conda
-  - pip install -r requirements.txt
   - export PATH=$HOME/conda/bin:$PATH
 
 install:
+  - pip install -r requirements.txt
   - python setup.py --cython install
 
 script: 
