### 2018-01-25 boisgera/bitstream#8 @bmcfee
I see.  Yeah, the revised doc clarifies this.  Thanks!
