### 2018-01-25 boisgera/bitstream#7 @bmcfee
How about just saying in the docs that the `State` object is meant for internal use only, and not meant for users?  That seems more direct than making a user follow a link and read an article, only to eventually discover that the class they're reading about isn't meant for them in the first place.
