### 2018-01-25 3b37939
install dev tools with requirements file (#4, #6)
