+++ .travis.yml
@@ -8,7 +8,7 @@ before_install:
   - export PATH=$HOME/conda/bin:$PATH
 
 install:
-  - pip install -r requirements.txt
+  - pip install -r requirements-dev.txt
   - python setup.py --cython install
 
 script: 
