### 2018-01-25 boisgera/bitstream#6 @boisgera
- I have added a "fat" `requirements-dev.txt` file that mainly targets the developers, and updated the documentation accordingly.

  - The mere user still needs to do this manually: we can't install a C compiler or pip via pip anyway, so at this point it's only a matter of installing numpy. And the user should know about numpy to use the library anyway ...

  - The user that wants to install from sources can either install "too many things" (all developer tools) easily or only Cython (manually).
