### 2018-01-25 boisgera/bitstream#7 @boisgera
But it is meant for users if they want to develop algorithms that "roll back" stream read or writes.
Arguably, this is an advanced pattern that most users won't need; and it only starts to make sense I guess if you look at this part of the API after having read the Snapshot section.
OTOH, its *implementation* is an internal detail.
I am dropping the link (yep, doesn't help much) and trying to explain this concisely.
