### 2018-01-25 7894734
new explanation for State (#7)
