### 2018-01-25 c0f345d
build docs (#7)
