### 2018-01-25 c32681c
motivation and link to applications in overview (#5)
