### 2018-01-25 boisgera/bitstream#5 @boisgera
Does it work for you? I have kept the motivation short in the overview, but the examples on the other hand are quite detailled.
