### 2018-01-29 boisgera/bitstream#5 @bmcfee
LGTM.
