### 2018-01-29 boisgera/bitstream#6 @bmcfee
This seems like a reasonable compromise to me.  Thanks for that!
