### 2018-01-29 boisgera/bitstream#10 @bmcfee
Seems reasonable.
