### 2018-01-29 boisgera/bitstream#7 @bmcfee
I guess my personal preference would be to document it anyway (for the sake of developers, including future-you), but it's not a hard preference, so I understand if you prefer to keep it this way.
