### 2018-01-29 #541 @bmcfee
@labarba The issues I raised above have all been sufficiently addressed by @boisgera -- I think it's in good shape now.
