### 2018-01-31 #541 @arfon
Hi @boisgera - could you please turn your bibliography into a bibtex file please? Our editorial tools aren't able to handle a JSON format.
