### 2018-01-31 78a3d94
convert bibliography to bibtex
