+++ paper/bibliography.bib
@@ -1,0 +1,68 @@
+
+@misc{bitstream,
+	title = {Bitstream: binary data for humans},
+	url = {https://github.com/boisgera/bitstream},
+	journal = {Github},
+	author = {Boisgérault, Sébastien},
+	year = {2017}
+}
+
+@misc{DAC,
+	title = {Digital {Audio} {Coding}},
+	url = {http://eul.ink/audio/},
+	journal = {Eul.Ink},
+	author = {Boisgérault, Sébastien},
+	year = {2017}
+}
+
+@misc{S1916,
+	title = {Digital {Audio} {Coding}},
+	url = {https://sgs.mines-paristech.fr/prod/sgs/ensmp/catalog/course/detail.php?code=S1916&year=3a&lang=EN},
+	journal = {MINES ParisTech Graduate School Course Catalog},
+	author = {{MINES ParisTech, PSL University}},
+	year = {2013}
+}
+
+@techreport{Rob94,
+	title = {{SHORTEN}: {Simple} lossless and near-lossless waveform compression},
+	url = {http://mi.eng.cam.ac.uk/reports/svr-ftp/auto-pdf/robinson_tr156.pdf},
+	author = {Robinson, Tony},
+	year = {1994}
+}
+
+@misc{struct,
+	title = {struct — {Interpret} strings as packed binary data},
+	url = {https://docs.python.org/2/library/struct.html},
+	author = {{Python 2.7 Standard Library}},
+	year = {2017}
+}
+
+@misc{array,
+	title = {array — {Efficient} arrays of numeric values},
+	url = {https://docs.python.org/2/library/array.html},
+	author = {{Python 2.7 Standard Library}},
+	year = {2017}
+}
+
+@misc{bitstring,
+	title = {Bitstring: a {Python} module to help you manage your bits},
+	url = {http://scott-griffiths.github.io/bitstring/},
+	journal = {Github},
+	author = {Griffiths, Scott},
+	year = {2014}
+}
+
+@misc{bitarray,
+	title = {Bitarray: efficient array of booleans for {Python}},
+	url = {https://github.com/ilanschnell/bitarray},
+	journal = {Github},
+	author = {Schnell, Ilan},
+	year = {2013}
+}
+
+@misc{WAVE,
+	title = {Audio {File} {Format} {Specification}},
+	journal = {MMSP Lab, ECE, McGill University},
+	author = {Kabal, Peter},
+	year = {2017}
+}
